Bonus/requirements.txt
# Bonus/requirements.txt
streamlit==1.33.0
numpy==1.26.4
Pillow==10.3.0
requests==2.32.3